requirements.txt
uvicorn 
fastapi
pymysql
python-dotenv
sqlalchemy
pydantic